1 AnyChatErrorCode.h 增加(视频呼叫-兼容早期的宏定义和GV_ERR_SUCCESS宏定义) 2 去掉ANYCHAT_INVALID_OBJECT_ID 宏定义的DWORD 3 将AnyChatQueuePlus ANYCHAT_QUEUE_INFO_LENGTH宏修改为ANYCHAT_QUEUE_INFO_QUEUELENGTH

+++ ios/sdk/include/AnyChatObjectDefine.h
@@ -16,7 +16,7 @@
 #define ANYCHAT_OBJECT_FLAGS_AUTOMODE	0x10	///< 自动服务模式
 #define ANYCHAT_OBJECT_FLAGS_GUESTLOGIN	0x20	///< 游客登录方式
 
-#define ANYCHAT_INVALID_OBJECT_ID	((DWORD)-1)	///< 无效的对象ID
+#define ANYCHAT_INVALID_OBJECT_ID           -1	///< 无效的对象ID
 
 // 坐席服务状态定义
 #define ANYCHAT_AGENT_STATUS_CLOSEED		0	///< 关闭，不对外提供服务（准备）
